feat(*): handle routing special cases
ref: MANAGER-17686
