+++ packages/manager-cli/routes/routes-migrate.mjs
@@ -1,7 +1,14 @@
 #!/usr/bin/env node
 
 import { execSync } from 'child_process';
+import { fileURLToPath } from 'url';
+import { dirname, resolve, join } from 'path';
 
+// Recreate __dirname
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const basePath = resolve('../manager/apps');
 const appName = process.argv[2];
 const isDryRun = process.argv.includes('--dry-run');
 const isValidAppName = /^[a-zA-Z0-9_-]+$/.test(appName);
@@ -13,19 +20,34 @@ if (!appName || !isValidAppName) {
   process.exit(1);
 }
 
-try {
-  // Build the command with optional dry-run flag
-  const transformCommand = `node ./routes/transform-routes-cli.mjs ${appName}${isDryRun ? ' --dry-run' : ''}`;
-  const updateCommand = `node ./routes/update-routers-init-cli.mjs ${appName}${isDryRun ? ' --dry-run' : ''}`;
+const migrateRoutes = async() => {
+  try {
+    // Build the command with optional dry-run flag
+    const transformCommand = `node ./routes/transform-routes-cli.mjs ${appName}${isDryRun ? ' --dry-run' : ''}`;
+    const updateCommand = `node ./routes/update-routers-init-cli.mjs ${appName}${isDryRun ? ' --dry-run' : ''}`;
 
-  console.log(`Running transform-routes-cli for ${appName}...`);
-  execSync(transformCommand, { stdio: 'inherit' });
+    console.log(`🧪Running transform-routes-cli for ${appName}...`);
+    execSync(transformCommand, { stdio: 'inherit' });
 
-  console.log(`Running update-routers-init-cli for ${appName}...`);
-  execSync(updateCommand, { stdio: 'inherit' });
+    console.log(`🧪Running update-routers-init-cli for ${appName}...`);
+    execSync(updateCommand, { stdio: 'inherit' });
 
-  console.log('Migration completed successfully');
-} catch (error) {
-  console.error('❌ Migration failed:', error);
-  process.exit(1);
+    // format changed code
+    const appFullPath = join('packages', 'manager', 'apps', appName, '**','*.tsx')
+    execSync(`yarn prettier --write ${appFullPath}`, {
+      stdio: 'inherit',
+      cwd: resolve(__dirname, '../..'),
+    })
+    execSync(`yarn eslint ${appFullPath} --fix --quiet`, {
+      stdio: 'inherit',
+      cwd: resolve(__dirname, '../..'),
+    })
+
+    console.log('✅ Migration completed successfully');
+  } catch (error) {
+    console.error('❌ Migration failed:', error);
+    process.exit(1);
+  }
 }
+
+migrateRoutes().catch((err) => console.error('❌ Unexpected error:', err));
