+++ packages/manager-cli/routes/update-routers-init-cli.mjs
@@ -119,9 +119,13 @@ const updateRouterInitialization = (code) => {
     },
   );
 
+  transformed = transformed.replace(
+    /import\s+\{\s*Routes\s*\}\s+from\s+['"](@\/|(\.{1,2}\/)+)routes\/routes['"];?\n?/g,
+    ''
+  );
+
   // Add `import Routes from './routes/routes'` if missing
-  if (!/import\s+Routes\s+from\s+['"][^'"]+routes['"]/.test(transformed)
-        && !transformed.includes('/routes/routes')) {
+  if (!/import\s+Routes\s+from\s+['"][^'"]+routes['"]/.test(transformed)) {
     transformed = `import Routes from '@/routes/routes';\n` + transformed;
   }
